Implemented the shuffle feature.
Shuffle image button in PlaybackFragment turns black when enabled.

The shuffle is implemented through the use of a new ArrayList in MediaPlayerManager. When suffle is enabled, this new ArrayList keeps the unshuffled queue, and the queue arraylist is shuffled. When unshuffled, new songs are added to this arraylist and removed songs are removed, and then this becomes the new queue.